added logic for address removal on Accounts
Purpose of change is to handle removing of address in ereg- until now when both for eks. postadresse and forretningsadresse were filled in ereg and afterwards Employe removed one of those addresses in ereg, Account in salesforce was not updated- the address was not removed. Added logic now to handle removal of addresses.